# openapi.yaml
# Generated from the running FastAPI app by scripts/generate_openapi.py.
# ORG_STANDARDS #6 requires this at the repo root. CI regenerates it and
# fails if it differs from what is committed, so it cannot drift from the
# endpoints that actually exist.
components:
  schemas:
    ExecuteRequest:
      properties:
        body:
          title: Body
          type: string
        branch:
          title: Branch
          type: string
        message:
          title: Message
          type: string
        path:
          title: Path
          type: string
      required:
      - path
      - body
      - message
      - branch
      title: ExecuteRequest
      type: object
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
    RunRequest:
      properties:
        approve:
          default: false
          title: Approve
          type: boolean
        pr:
          default: 4471
          title: Pr
          type: integer
        seed:
          default: false
          title: Seed
          type: boolean
      title: RunRequest
      type: object
    ValidationError:
      properties:
        ctx:
          title: Context
          type: object
        input:
          title: Input
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
info:
  description: One image, three deployments. What differs is the service account Cloud
    Run starts it with and MITOS_ROLE, neither of which the process can change. /execute
    exists on all three deployments and refuses on two of them.
  license:
    name: MIT
    url: https://github.com/upgradedev/mitos-gcp/blob/main/LICENSE
  title: Mitos fleet API
  version: 1.0.0
openapi: 3.1.0
paths:
  /:
    get:
      operationId: index__get
      responses:
        '200':
          content:
            text/html:
              schema:
                type: string
          description: Successful Response
      summary: Index
  /catalog:
    get:
      operationId: catalog_catalog_get
      responses:
        '200':
          content:
            application/json:
              schema:
                additionalProperties: true
                title: Response Catalog Catalog Get
                type: object
          description: Successful Response
      summary: Catalog
  /execute:
    post:
      description: 'Perform the governed write. Writer service only.


        The role check here is not decoration: this endpoint exists on all three

        deployments because they share one image, and on two of them it refuses.

        Even if it did not, those two cannot read the deploy key.'
      operationId: execute_execute_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExecuteRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                additionalProperties: true
                title: Response Execute Execute Post
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Execute
  /identity:
    get:
      operationId: identity_identity_get
      responses:
        '200':
          content:
            application/json:
              schema:
                additionalProperties: true
                title: Response Identity Identity Get
                type: object
          description: Successful Response
      summary: Identity
  /run:
    post:
      operationId: run_run_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RunRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Run
  /run/stream:
    post:
      description: 'The chore, streamed as it happens.


        `/run` returns when the whole thing is finished, and with a model in the

        loop that is minutes rather than seconds: the specialists read the

        repository, and each read is a round trip. A judge who posts to it waits,

        sees nothing, and concludes it is broken. The UAT caught exactly that.


        So this streams. Every beat is flushed the moment it occurs, which is both

        the honest fix for the timeout and a better thing to watch: the reads appear

        one at a time, in the order the agent chose them.'
      operationId: run_stream_run_stream_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RunRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Run Stream
  /thread:
    get:
      operationId: thread_thread_get
      parameters:
      - in: query
        name: limit
        required: false
        schema:
          default: 100
          title: Limit
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                additionalProperties: true
                title: Response Thread Thread Get
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Thread
  /thread/view:
    get:
      description: 'The thread as the graph it is.


        The product is named for a thread you can follow back, and rendering it as

        a list asks the reader to do the walking in their head. Here a click lights

        the whole path from an outcome to the pull request that caused it.'
      operationId: thread_view_thread_view_get
      parameters:
      - in: query
        name: limit
        required: false
        schema:
          default: 300
          title: Limit
          type: integer
      responses:
        '200':
          content:
            text/html:
              schema:
                type: string
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Thread View
  /watch:
    get:
      description: 'Proof the subscription is real, and countable.


        `wakeups` only increments when Firestore delivered a snapshot in which a

        deferral had expired. Nothing in this codebase calls the fleet to produce

        one.'
      operationId: watch_watch_get
      responses:
        '200':
          content:
            application/json:
              schema:
                additionalProperties: true
                title: Response Watch Watch Get
                type: object
          description: Successful Response
      summary: Watch
  /webhook/github:
    post:
      description: 'The trigger. Nobody opens Mitos; this does.


        Answers immediately and works afterwards. GitHub gives a webhook ten seconds

        and disables one that keeps timing out, and this chore takes minutes with
        a

        model reading the repository, so doing the work inside the request would

        guarantee the trigger stops firing.'
      operationId: github_webhook_webhook_github_post
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
      summary: Github Webhook
servers:
- description: reader, holds the Firestore query subscription
  url: https://mitos-reader-437828525303.europe-west1.run.app
- description: evaluator
  url: https://mitos-evaluator-437828525303.europe-west1.run.app
- description: writer, the only identity that can publish
  url: https://mitos-writer-437828525303.europe-west1.run.app